refactor Trails to allow distance float

+++ src/main/java/dao/TrailDao.java
@@ -15,7 +15,7 @@ public interface TrailDao {
 
     Trail findById(int id); // Find Trail by id
 
-    void update(int id, String name, String difficulty, String location, Double latitude, Double longitude, Integer distance);//update
+    void update(int id, String name, String difficulty, String location, Double latitude, Double longitude, Double distance);//update
 
     void deleteById(int id); //Delete
